Moved functions analyzing message data to new file. Added documentation in .md files
