+++ .sonarqube/sonar-project.properties
@@ -12,5 +12,6 @@ src/carma-platform/run_all_control_analysis.py, \
 src/carma-platform/environment_scripts.py, \
 src/carma-platform/carma_system_resource_usage_plot.py, \
 src/carma-platform/run_all_workzone_analysis.py, \
+src/carma-platform/message_scripts.py, \
 src/carma-platform/run_all_regression_analysis.py
 src/carma-platform/utils.py
